+++ SparseGrids/tsgGridGlobal.hpp
@@ -135,6 +135,13 @@ protected:
     void loadConstructedTensors();
     std::vector<int> getMultiIndex(const double x[]);
 
+    #ifdef Tasmanian_ENABLE_CUDA
+    void loadCudaValues() const;
+    void clearCudaValues() const;
+    void loadCudaNodes() const;
+    void clearCudaNodes() const;
+    #endif
+
 private:
     TypeOneDRule rule;
     double alpha, beta;
@@ -158,7 +165,7 @@ private:
     std::unique_ptr<DynamicConstructorDataGlobal> dynamic_values;
 
     #ifdef Tasmanian_ENABLE_CUDA
-    mutable CudaVector<double> cuda_values;
+    mutable std::unique_ptr<CudaGlobalData<double>> cuda_cache;
     #endif
 };
 #endif // __TASMANIAN_DOXYGEN_SKIP